Fix E2E bundle freshness (Fixes #2240)
