+++ packages/cli/src/ui/hooks/geminiStream/useGeminiStream.ts
@@ -17,7 +17,6 @@ import {
   type Config,
   type AgentClientContract,
   type EditorType,
-  type ThinkingBlock,
   type MessageBus,
   type RecordingIntegration,
 } from '@vybestack/llxprt-code-core';
@@ -34,56 +33,6 @@ import { useCheckpointPersistence } from './checkpointPersistence.js';
 import type { useStreamState } from './useStreamState.js';
 import type { GeminiStreamOrchestrationDeps } from './useGeminiStreamOrchestration.js';
 import { useGeminiStreamOrchestration } from './useGeminiStreamOrchestration.js';
-
-/**
- * Resets or carries-over per-turn state depending on whether this is a new
- * prompt or a continuation. Also handles bucket failover reset/reauth.
- */
-function bindOptionalConfigMethod<T extends (...args: never[]) => unknown>(
-  method: T | undefined,
-  config: Config,
-): T | undefined {
-  return method?.bind(config) as T | undefined;
-}
-
-export async function prepareTurnForQuery(
-  isContinuation: boolean,
-  config: Config,
-  startNewPrompt: () => void,
-  setThought: (t: null) => void,
-  thinkingBlocksRef: React.MutableRefObject<ThinkingBlock[]>,
-): Promise<void> {
-  const getBucketFailoverHandler = bindOptionalConfigMethod(
-    config.getBucketFailoverHandler,
-    config,
-  );
-
-  if (!isContinuation) {
-    startNewPrompt();
-    setThought(null);
-    thinkingBlocksRef.current = [];
-    getBucketFailoverHandler?.()?.reset?.();
-
-    // Invalidate auth cache at turn boundaries for new turns
-    // This ensures tokens updated by other processes are picked up
-    const handler = getBucketFailoverHandler?.();
-    if (handler?.invalidateAuthCache) {
-      const getRuntimeSessionId = bindOptionalConfigMethod(
-        config.getSessionId,
-        config,
-      );
-      const runtimeId = getRuntimeSessionId?.() ?? 'default';
-      handler.invalidateAuthCache(runtimeId);
-    }
-  } else {
-    getBucketFailoverHandler?.()?.resetSession?.();
-  }
-  try {
-    await getBucketFailoverHandler?.()?.ensureBucketsAuthenticated?.();
-  } catch {
-    // Swallow — partial auth is acceptable.
-  }
-}
 
 export const useGeminiStream = (
   agentClient: AgentClientContract,
